CryptoAI/requirements.txt
# Core Data Fetching (FREE APIs - No Pro Tier Required!)
requests==2.31.0
ccxt==4.2.25  # For Coinbase Advanced Trade & other exchanges
python-dotenv==1.0.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.26.2
ta==0.11.0  # Technical Analysis library

# Authentication & Security
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
cryptography==41.0.7
ta-lib==0.4.28  # Advanced technical indicators

# Visualization
matplotlib==3.8.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Web & API
flask==3.0.0
websocket-client==1.7.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
schedule==1.2.1

# Database (optional for storing history)
sqlalchemy==2.0.23

# Machine Learning & Prediction
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.16.2
keras==3.13.1
lightgbm==4.1.0  # Alternative to XGBoost
prophet==1.1.5  # Facebook's time series forecasting
